Struct file_tree::FileTree

pub struct FileTree { /* fields omitted */ }

Creates a directory structure suitable for storing large numbers of files. Optionally deletes the created directory and files when dropped.

Methods

impl FileTree

pub fn new_in(path: PathBuf, persistent: bool) -> Result<FileTree>

Create a new directory structure under path. If persistent is false the directory and all it's contents will be deleted when the returned FileTree is dropped

Errors

If persistent is false, the directory will be created using tempdir::TempDir, and any related errors will be returned here

pub fn new(persistent: bool) -> Result<FileTree>

Create a new directory structure. If persistent is false the directory and all it's contents will be deleted when the returned FileTree is dropped.

Errors

If persistent is false, the directory will be created using tempdir::TempDir, and any related errors will be returned here

pub fn get_new_file(&mut self) -> Result<PathBuf>

Returns a PathBuf pointing to an available slot in the file tree. The file pointed to by the returned PathBuf will not be created by this method call.

Errors

If a new subdirectory is required, fs::create_dir_all will be called. Any errors from that call will be returned here

pub fn get_path(&self) -> PathBuf

Return the root path for the file tree